First Cut at Expression Parsing
Adds a simple parser which turns a single value token into a value
expression. Now needs work on the `Parser::parse()` method to allow
parsing of full expressions via precedence parsing.

+++ src/expression.rs
@@ -19,7 +19,7 @@ pub enum Expression<'a> {
     /// Value Expression
     ///
     /// Represents a simple numeric value
-    ValueExpression(i32),
+    ValueExpression(i64),
 
     /// Variable Expression
     ///
